pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cascadeui"
version = "2.0.0"
description = "Redux-inspired UI framework for discord.py"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "HollowTheSilver", email = "hollow@users.noreply.github.com" }
]
keywords = ["discord", "discord.py", "ui", "components", "state-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]
dependencies = [
    "discord.py>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "isort>=5.12",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "pymdown-extensions>=10.0",
]
sqlite = [
    "aiosqlite>=0.19",
]
redis = [
    "redis>=5.0",
]

[project.urls]
Homepage = "https://github.com/HollowTheSilver/CascadeUI"
Repository = "https://github.com/HollowTheSilver/CascadeUI"
Issues = "https://github.com/HollowTheSilver/CascadeUI/issues"
Documentation = "https://hollowthesilver.github.io/CascadeUI/"

[tool.setuptools.packages.find]
include = ["cascadeui*"]

[tool.setuptools.package-data]
cascadeui = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
